Spaced Repetition
This program implements spaced-repetition to help with memorizing
things.  Point the program to a data file and it'll process through
all the lines.  Using text, since that is a universal interface.

Question the user and print out the results of the next file

Learn about the double-getchar idiom.  At some point we might want to
switch to something better like fgets.

This program now reads from a file and writes to a file

The output file can be the same as the input file.  I could force that
the program to always write back into the same "database" file, but
I'd just rather allow the user to use an immutable approach.  If the
user doesn't want that, they can always just output back to the same
file.